init the sse_inference
