+++ Cargo.toml
@@ -24,3 +24,6 @@ once_cell = "1.17"
 prometheus = "0.13"
 argh = "0.1"
 axum = "0.6.15"
+futures = "0.3.28"
+futures-util = "0.3.28"
+async-stream = "0.3.5"
